separate function from utils that were actually dedicated to geometrical operation and lines operation
