+++ colibrie/geometry_operation.py
@@ -1,0 +1,32 @@
+from colibrie.geometry import Point
+
+
+def intersect(p1, p2, p3, p4):
+    x1, y1 = p1
+    x2, y2 = p2
+    x3, y3 = p3
+    x4, y4 = p4
+    denom = (y4 - y3) * (x2 - x1) - (x4 - x3) * (y2 - y1)
+    if denom == 0:  # parallel
+        return None
+    ua = ((x4 - x3) * (y1 - y3) - (y4 - y3) * (x1 - x3)) / denom
+    if ua < 0 or ua > 1:  # out of range
+        return None
+    ub = ((x2 - x1) * (y1 - y3) - (y2 - y1) * (x1 - x3)) / denom
+    if ub < 0 or ub > 1:  # out of range
+        return None
+    x = x1 + ua * (x2 - x1)
+    y = y1 + ua * (y2 - y1)
+    return Point(x, y)
+
+
+def is_rectangle_overlap(r1, r2):
+    if (r1[0] >= r2[2]) or (r1[2] <= r2[0]) or (r1[3] <= r2[1]) or (r1[1] >= r2[3]):
+        return False
+    else:
+        return True
+
+
+def is_rectangle_contained(r1, r2):
+    return r1[0] <= r2[0] <= r2[2] <= r1[2] and r1[1] <= r2[1] <= r2[3] <= r1[3]
+
